cipher
With this code, you should be able to paste any block of text on line 27 (or between the triple parenthesis) and change the number on line 32 in order to encode the block of text with a caesar cipher, according to the number you chose to input. It removes all spaces and punctuation, as well as changing all capitals to lower case. This produces a long string of text, which makes it extra confusing (because a cipher should be!)

In this instance, I encoded a block of text at an interval of +10, which should process as the following:
"
drscmycwsmnkxmoyplebcdsxqnomknoxmokxngsdrrovnzobwsccsyxcdgscdckvvyebkbwcmyvvomdsfoviledspcgoodxoccmkxgsxkxnsdmkxdroxsvvcdsvvlorobodywybbygdyrsqrpsfoiyeiocdobnkiwipbsoxn
"

You can use this same code to decode a caesar-ciphered message by pasting the block of encoded text between the triple parenthesis and changing the number on line 32 to the negative of the number which was used to encode it in the first place. 
In this case, you would decode by inputting "-10" because I initially encoded it at an interval of +10.